[feat]프로젝트 상세 조회 기능 구현

+++ src/main/java/com/nexus/sion/feature/project/query/dto/request/ProjectListRequest.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class ProjectListRequest {
   private String keyword;
   private Long maxBudget;
-  private Double maxPeriodInMonth;
+  private Integer maxPeriodInMonth;
   private Integer maxNumberOfMembers;
   private List<String> statuses;
